requirements.txt
# Django and core dependencies
Django==4.2.13  # LTS version for production stability
django-environ==0.11.2
django-allauth==0.57.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0  # JWT Authentication
django-cryptography==1.1  # Compatible with Django 4.2
django-ratelimit==4.1.0  # Rate limiting

# Database - SQLite only (built into Python)
# No additional database drivers needed

# Forms and UI
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-htmx==1.17.0

# File handling and PDF generation
WeasyPrint==60.2
Pillow>=10.0.0,<11.0.0
python-magic>=0.4.24

# WSGI server for production
gunicorn==21.2.0

# Authentication and security
django-otp==1.2.0
django-formtools==2.5.1
argon2-cffi==25.1.0  # Enhanced password hashing

# Cache backends
django-redis==5.4.0
redis>=4.5.0,<6.0.0
hiredis==2.3.2

# Performance and monitoring
django-extensions==3.2.3
psutil>=5.9.0
requests>=2.31.0
sentry-sdk[django]==1.40.6

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0

# Production monitoring
setproctitle>=1.3.0  # Better process naming for monitoring

# Development and testing
django-debug-toolbar==4.2.0  # Only for development

# Testing dependencies
coverage>=7.0.0
factory-boy>=3.2.0
freezegun>=1.2.0
mock>=5.0.0
responses>=0.23.0
model-bakery>=1.15.0

# API Documentation
drf-yasg==1.21.10
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5
